+++ extract_msg/named.py
@@ -1,5 +1,6 @@
 import copy
 import logging
+import pprint
 import zlib
 
 from extract_msg import constants