+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Added support for `PtypFloatingTime` in `utils.parseType`.
 * Added proper support for `PtypTime` in `FixedLengthProp.parseType`
 * Added pretty print functions to the `Properties` class and the `Named` class.
+* Added new function `MSGFile._ensureSetProperty` that acts like `MSGFile._ensureSet` except that it works with properties from the `Properties` instance.
+* Added equivalent of the previously mentioned function to the `Attachment` class and the `Recipient` class.
 
 **v0.26.4**
 * Added new function `MSGFile._ensureSetNamed` which acts like `MSGFile._ensureSet` except that it works with named properties.